Update README.md
The Project

Akitacoin is a cryptocurrency that does not have any utility or intrinsic value. It is associated with internet memes, jokes
								
Akitacoin is inspired by Dogecoin, but with a couple of key differences. Firstly, the BNB Smart Chain blockchain was chosen for its placement, and its own was not developed. Secondly, the emission is limited to 357,888,888,888,888 tokens. In addition, Akitacoin expects to become a more sustainable project in the long term through the launch of additional products.